Este commit contiene las relaciones, y el archivo de bd de valentinos

+++ database/migrations/2023_03_09_022944_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('url');
-            $table->decimal('price', 8, 2);
+            $table->float('price');
             $table->string('combo_ids');
             $table->unsignedBigInteger('product_type_id');
             $table->boolean('status');
